--- cor/hardware/avr/0.0.1/bootloaders/bootloader.txt
@@ -1,1 +1,0 @@
-
